♻️ refactor :: export 하는 함수들과 파일명 일치하도록 리팩토링
